fix(eqnedit64): make an unmapped embellishment a build failure, not a hat
Adding a test for the prime-drawn-as-a-hat defect left the shape that caused
it: an if-chain whose fall-through was the hat, so the next embellishment
anyone adds becomes a hat again and no one is told. That is a silent fallback
in the sense the No Fallbacks policy means - a plausible wrong answer where an
error belongs - and three surfaces each had their own version of it.

Now the compiler refuses it. `embellType` and `embells` carry the
Embellishment enum rather than int, the renderer and the MathML emitter switch
over it with no default, and the build enables C4062 - MSVC leaves that
warning off even at /W4, which is why an unmapped enumerator compiled cleanly
when first tried. With /WX an embellishment nobody mapped now stops the build.

The check found a defect on its first run: the MathML emitter published
neither harpoon accent, so an equation carrying one lost it on the way to
Office - the same class as the drawing bug, in the other direction. Both are
published now.

Verified the guard is awake rather than assumed: adding an enumerator makes
both switches fail the build with C4062, and removing it builds clean again.
A test can only catch what someone thought to write down; this cannot be
forgotten, because nothing compiles until the question is answered.

+++ tools/eqnedit64/src/latex_emitter.cpp
@@ -882,7 +882,7 @@ void LaTeXEmitter::emitMatrix(const MatrixNode& mat, std::string& out) {
 
 void LaTeXEmitter::emitEmbell(const EmbellNode& embell, std::string& out) {
     const std::string content = emitNodes(embell.content);
-    const int type = embell.embellType;
+    const int type = static_cast<int>(embell.embellType);
     /* The prime family has no prefix -- it is written entirely as a suffix
      * (x', x'').  Requiring both halves silently dropped the mark, so a
      * prime template produced a bare x. */
